api/pyproject.toml
[tool.poetry]
name = "loramon-api"
version = "0.1.0"
description = "API FastAPI para LoraMon"
authors = ["Você <arturcarvalho@grpereira.com.br>"]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
psycopg = {version = "^3.2.1", extras=["binary"]}
pydantic = "^2.8.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
PyMuPDF = "^1.24.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"
